Move Dashboard component files into components/Dashboard
